requirements.txt
trimesh[easy]
pyrender
matplotlib
pillow
numpy
tqdm
